refactor: 라우터에서 도메인 로직 분리
